+++ src/lib/appState/main/operations.ts
@@ -1,69 +1,6 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
 import { apiIngco, serializeAxiosError } from '../user/operation';
-import { sortValueType } from '@/app/ui/catalog/FiltersBlock';
-import { normalizeProduct, normalizeOrder } from '@/lib/utils';
-
-
-export const fetchMainTableDataThunk = createAsyncThunk(
-  'mainTable/fetch',
-  async (
-    {
-      page,
-      query,
-      category,
-      limit = 30,
-      sortValue,
-      isRetail = true,
-      filters,
-    }: {
-      query: string;
-      page: number;
-      category?: string;
-      limit?: number;
-      sortValue: sortValueType;
-      isRetail: boolean;
-      filters?: string;
-    },
-    { rejectWithValue, signal },
-  ) => {
-    try {
-      const { data } = await apiIngco.get('/products', {
-        params: { page, q: query, limit, category, sortValue, isRetail, filters },
-        signal,
-      });
-      return { ...data, products: data.products.map(normalizeProduct) };
-    } catch (error) {
-      return rejectWithValue(serializeAxiosError(error));
-    }
-  },
-);
-
-export const getProductByIdThunk = createAsyncThunk(
-  'product/fetch',
-  async (productId: string, { rejectWithValue }) => {
-    try {
-      const { data } = await apiIngco.get(`/products/id/${productId}`);
-      return normalizeProduct(data);
-    } catch (error) {
-      return rejectWithValue(serializeAxiosError(error));
-    }
-  },
-);
-
-export const getProductBySlugThunk = createAsyncThunk(
-  'product/fetchBySlug',
-  async (productSlug: string, { rejectWithValue }) => {
-    try {
-      const { data } = await apiIngco.get(`/products/${productSlug}`);
-      return normalizeProduct(data);
-    } catch (error) {
-      return rejectWithValue(serializeAxiosError(error));
-    }
-  },
-);
-
-
-
+import { normalizeOrder } from '@/lib/utils';
 
 export const fetchHistoryThunk = createAsyncThunk(
   'history/fetch',
@@ -90,21 +27,6 @@ export const fetchHistoryThunk = createAsyncThunk(
     }
   },
 );
-
-export const deleteProductThunk = createAsyncThunk(
-  'product/delete',
-  async (productId: number, { rejectWithValue }) => {
-    try {
-      await apiIngco.delete(`/products/${productId}`);
-      return productId;
-    } catch (error) {
-      return rejectWithValue(serializeAxiosError(error));
-    }
-  },
-);
-
-
-
 
 export const fetchExcelFileThunk = createAsyncThunk(
   'excel/fetch',